apps/api/pyproject.toml
[project]
name = "eventcart-api"
version = "0.1.0"
description = "EventCart FastAPI backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.30",
  "sqlalchemy>=2.0",
  "asyncpg>=0.29",
  "psycopg2-binary>=2.9",
  "alembic>=1.13",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "email-validator>=2.1",
  "python-jose[cryptography]>=3.3",
  "argon2-cffi>=23.1",
  "passlib>=1.7",
  "structlog>=24.1",
  "python-multipart>=0.0.9",
  "orjson>=3.10",
]

[project.optional-dependencies]
 dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "httpx>=0.27",
  "ruff>=0.4",
  "mypy>=1.10",
  "types-python-jose>=3.3.4.20240106",
]

[tool.uv]
package = true

[tool.ruff]
line-length = 100
src = ["src"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.12"
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
